ddt: emmc/mmc: Add stress pattern testing
This adds initial support for stress pattern testing in blk_device_dd_readwrite_test.
Variable like PATTERN{1-9} shall hold pattern types used for stress pattern
testing. Script decides to do stress pattern testing instead of random data testing
by -p flag passed to the script, where -p "1" is taken by 0xff00 stress pattern, next
stress pattern -p "2" shall be next stress pattern added in the future.

+++ testcases/ddt/scripts/blk/blk_device_dd_readwrite_test.sh
@@ -20,6 +20,8 @@ source "mtd_common.sh"
 
 SKIP_FORMAT=0
 WRITE_TO_FILL=0
+STRESS_PATTERN="0"
+PATTERN1="\x00\xff\x00\xff\x00\xff\x00\xff"
 ############################# Functions #######################################
 usage()
 {
@@ -30,6 +32,7 @@ cat <<-EOF >&2
   -m MNT_POINT    mount point 
   -b DD_BUFSIZE   dd buffer size for 'bs'
   -c DD_CNT       dd count for 'count'
+  -p STRESS_PATTERN      Use specific stress pattern instead of random data
   -i IO_OPERATION IO operation like 'wr', 'cp', default is 'wr'
                   'oversize_write' is to test if driver throw error when the size > partition size
   -d DEVICE_TYPE  device type like 'nand', 'mmc', 'usb' etc
@@ -53,7 +56,7 @@ compare_md5sum()
 }
 ############################### CLI Params ###################################
 
-while getopts  :d:f:m:n:b:c:i:l:swh arg
+while getopts  :d:f:m:n:b:c:p:i:l:swh arg
 do case $arg in
         n)      
                 # optional param
@@ -63,6 +66,7 @@ do case $arg in
         m)      MNT_POINT="$OPTARG"; MNT_POINT="${MNT_POINT}_$$";;
         b)      DD_BUFSIZE="$OPTARG";;
         c)      DD_CNT="$OPTARG";;
+        p)      STRESS_PATTERN="$OPTARG";;
         i)      IO_OPERATION="$OPTARG";;
         l)      TEST_LOOP="$OPTARG";;
         s)      SKIP_FORMAT=1;;
@@ -156,7 +160,38 @@ test_print_trc "Doing read/write test for $TEST_LOOP times"
 # not using tmpfs because it is too small and we don't measure performance here
 #SRC_FILE='/dev/shm/srctest_file' 
 SRC_FILE="$HOME/srctest_file_${DEVICE_TYPE}_$$"
-do_cmd "time dd if=/dev/urandom of=$SRC_FILE bs=$DD_BUFSIZE count=$DD_CNT"
+
+if [ "$STRESS_PATTERN" != "0" ]; then
+  BSIZE="0"
+  echo "Debug: Using stress pattern=$STRESS_PATTERN"
+  case "$DD_BUFSIZE" in
+    "1" | "512")
+      BSIZE="$DD_BUFSIZE";;
+    "64K")
+      BSIZE="64000";;
+    "512K")
+      BSIZE="512000";;
+    "5M")
+      BSIZE="5000000";;
+    "10M")
+      BSIZE="10000000";;
+  esac
+
+  if [ "$STRESS_PATTERN" != "1" ] || [ $BSIZE == "0" ]; then
+    echo "Not a supported blocksize for stress pattern testing, revert to random data"
+    do_cmd "time dd if=/dev/urandom of=$SRC_FILE bs=$DD_BUFSIZE count=$DD_CNT"
+  else
+    case "$STRESS_PATTERN" in
+      "1")
+        size=$((BSIZE * DD_CNT / 8))
+        for ((i=0; i<$size; i++)); do printf $PATTERN1 >> $SRC_FILE; done
+      ;;
+    esac
+  fi
+else
+  do_cmd "time dd if=/dev/urandom of=$SRC_FILE bs=$DD_BUFSIZE count=$DD_CNT"
+fi
+
 sleep 10
 do_cmd ls -lh $SRC_FILE
 do_cmd "df -h"
